Organize project java classes by clean architecture design
Application is centered around domain core, which contains the model
and use cases. These define the interfaces for the services and
persistence layers.

The application plain java services are wired-up to the application
spring context in the @Configuration classes inside the config
package.

Persistence is split in dao objects and spring repository interfaces


Former-commit-id: 3e234e51586f8e8ca63878cb62446d834409d1f8