Hash MathML markup and use it as filename for server storage
MathML library return an xml markup of the math equations and
an image that is sent to back-end; There is an issue where
the back-end doesn't save the images properly because it
tries to hash them. Something goes wrong there and the
url that is sent back is from another mathml image.

The hashing of the back-end is controlled via a query param
`isSha1` which is always set to true. I added the flexibility
to change it to false and save the file using the fileName provided.

Fixing Bug 1433


Former-commit-id: 95e9b728366ee692c8179c797b1a7ab80943c0fb